Touch-up for internal BaseDataStore:InternalQuery to simplify deletion cases
